requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.2
anyio==4.9.0
autograd==1.8.0
autoray==0.7.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
flask==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
igraph==0.11.8
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
mcp==1.8.1
numpy==2.2.5
pycryptodome==3.22.0
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
python-constraint==1.4.0
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
scipy==1.15.3
sniffio==1.3.1
sse-starlette==2.3.5
starlette==0.46.2
texttable==1.7.0
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.2
werkzeug==3.1.3
